pyproject.toml
[project]
name = "review-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.6.1",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "ipykernel>=7.0.1",
    "langchain-community>=0.3.31",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "notebook>=7.4.7",
    "rich>=14.2.0",
    "fastmcp>=2.12.4",
    "langchain-mcp-adapters>=0.1.11",
]
